Resolver os testes de 'EquipmentStateHistory'
- Código testado a nível cliente.

+++ BusOnTime.Application/Mapping/DTOs/InputModel/EquipmentPositionHistoryIM.cs
@@ -11,7 +11,7 @@ namespace BusOnTime.Application.Mapping.DTOs.InputModel
     {
         public Guid? EquipmentId { get; set; }
         public DateTime Date { get; set; }
-        public int Lat { get; set; }
-        public int Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
     }
 }
