+++ BusOnTime.Data/Entities/EquipmentPositionHistory.cs
@@ -16,8 +16,8 @@ namespace BusOnTime.Data.Entities
         public Guid EquipmentPositionId { get; set; }
         public Guid? EquipmentId { get; set; }
         public DateTime Date { get; set; }
-        public int Lat { get; set; }
-        public int Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
         public Equipment? Equipment { get; set; }
     }
 }
